ideas routs update
