+++ index.js
@@ -2,6 +2,7 @@ const express = require('express');
 const cors = require('cors');
 const md5 = require('md5');
 const cookieParser = require('cookie-parser');
+const ideasRoute = require('./routes/ideas');
 const {
     login,
     registerUser,
@@ -17,15 +18,6 @@ const {
     createDonationByRegisteredUser,
 } = require('./businessRules/donations');
 const bodyParser = require('body-parser');
-const { updateIdea, deleteIdea, getIdeas, getIdea } = require('./models/ideas');
-const {
-    sortIdeasByTotalDonationSum,
-    getPendingIdeas,
-    getApprovedIdeas,
-    createNewIdea,
-    updateIdeasStatus,
-    getRejectedIdeas,
-} = require('./businessRules/ideas');
 
 const app = express();
 const port = process.env.PORT || 3003;
@@ -50,72 +42,7 @@ app.use(express.json());
 
 // IDEAS METHODS
 
-app.get('/ideas', (req, res) => {
-    let sortedIdeas = getIdeas();
-
-    if (req.query.sortBy === 'totalDonationSum') {
-        sortedIdeas = sortIdeasByTotalDonationSum(ideas);
-    } else if (
-        req.query.sortBy === 'status' &&
-        req.query.status === 'accepted'
-    ) {
-        sortedIdeas = getApprovedIdeas();
-    } else if (
-        req.query.sortBy === 'status' &&
-        req.query.status === 'pending'
-    ) {
-        sortedIdeas = getPendingIdeas();
-    } else if (
-        req.query.sortBy === 'status' &&
-        req.query.status === 'rejected'
-    ) {
-        sortedIdeas = getRejectedIdeas();
-    }
-
-    res.status(200).json(sortedIdeas);
-});
-
-app.post('/ideas', (req, res) => {
-    try {
-        createNewIdea(
-            req.body.header,
-            req.body.description,
-            req.body.askedSum,
-            req.body.userId
-        );
-    } catch (error) {
-        res.status(400).send(error.message);
-    }
-    res.status(200).send('Idea created successfully.');
-});
-
-app.put('/ideas', (req, res) => {
-    try {
-        updateIdea(req.body.ideaId, req.body.key, req.body.value);
-    } catch (error) {
-        res.status(404).send(error.message);
-    }
-    res.status(200).send('Idea updated successfully.');
-});
-
-app.put('/ideas/status', (req, res) => {
-    try {
-        updateIdeasStatus(req.body.ideaId, req.body.isApproved);
-    } catch (error) {
-        res.status(404).send(error.message);
-    }
-    res.status(200).send('Ideas status updated successfully.');
-});
-
-app.delete('/ideas', (req, res) => {
-    try {
-        deleteIdea(req.body.ideaId);
-    } catch (error) {
-        res.status(404).send(error.message);
-    }
-
-    res.status(200).send('Ideas successfully deleted.');
-});
+app.use('/ideas', ideasRoute);
 
 //USER METHODS
 app.get('/users', (req, res) => {
